--- Example/demo.wiretapp/MockResponses/WireTappUITests/ExampleTest/:todos/0.json
@@ -1,23 +1,0 @@
-{
-    "status": 200,
-    "response": [
-        {
-          "userId": 1,
-          "id": 1,
-          "title": "delectus aut autem",
-          "completed": false
-        },
-        {
-          "userId": 1,
-          "id": 2,
-          "title": "quis ut nam facilis et officia qui",
-          "completed": false
-        },
-        {
-          "userId": 1,
-          "id": 3,
-          "title": "fugiat veniam minus",
-          "completed": false
-        }
-     ]
-}
