--- Example/demo.wiretapp/MockResponses/WireTappUITests/ExampleTest/:todos/1.json
@@ -1,23 +1,0 @@
-{
-    "status": 200,
-    "response": [
-        {
-          "userId": 1,
-          "id": 1,
-          "title": "Nick Gorman",
-          "completed": false
-        },
-        {
-          "userId": 1,
-          "id": 2,
-          "title": "Muhammad Ali",
-          "completed": false
-        },
-        {
-          "userId": 1,
-          "id": 3,
-          "title": "Md Rashidul",
-          "completed": false
-        }
-     ]
-}
