+++ Example/demo.wiretapp/MockResponses/default/:posts.json
@@ -1,0 +1,17 @@
+{
+  "status" : 200,
+  "response" : [
+    {
+      "body" : "body of post number 1, lorem ipsum est rerum tempore vitae\nsequi sint nihil reprehenderit dolor beatae ea dolores neque\nfugiat blanditiis voluptate porro vel nihil molestiae ut reiciendis\nqui aperiam non ",
+      "id" : 1,
+      "title" : "a great title for the first post",
+      "userId" : 1
+    },
+    {
+      "body" : "est rerum tempore vitae\nsequi sint nihil reprehenderit dolor beatae ea dolores neque\nfugiat blanditiis voluptate porro vel nihil molestiae ut reiciendis\nqui aperiam non debitis possimus qui neque nisi nulla",
+      "id" : 2,
+      "title" : "qui est esse",
+      "userId" : 1
+    }
+  ]
+}
