+++ Example/demo.wiretapp/MockResponses/default/:users.json
@@ -1,0 +1,28 @@
+{
+  "status" : 200,
+  "response" : [
+    {
+      "address" : {
+        "geo" : {
+          "lat" : "-37.3159",
+          "lng" : "81.1496"
+        },
+        "street" : "East Liberty Street",
+        "suite" : "Apt. 556",
+        "city" : "Gwenborough",
+        "zipcode" : "92998-3874"
+      },
+      "phone" : "1-770-736-8031 x56442",
+      "website" : "hildegard.org",
+      "id" : 1,
+      "company" : {
+        "name" : "Romaguera-Crona",
+        "catchPhrase" : "Multi-layered client-server neural-net",
+        "bs" : "harness real-time e-markets"
+      },
+      "username" : "NickG",
+      "email" : "Sincere@april.biz",
+      "name" : "Nick Gorman"
+    }
+  ]
+}
