+++ Example/demo.wiretapp/demo.wiretapp/Views/WideText.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+struct WideText: View {
+    let text: String
+    var body: some View {
+        Text(text)
+            .font(.caption)
+            .frame(
+                maxWidth: .infinity,
+                alignment: .leading
+            )
+    }
+}
